+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Added
+- Resources may now be searched by name from the entity manager using the `search` query field.
+
 ### Fixed
 - Timezones were not being saved correctly for resources.